eja mater: show all bar numbers ..
I'm not sure whether the default setting (showing bar numbers
only at the beginning of the line) is good - personally I
think it's better to have all of them visible in a piece like this.
Anyway, I'm turning them on because I want to compare LilyPond's
engraving to another one which has all bar numbers shown.