eja mater: make lines thinner - fixme level 1
Thinner lines result in notation elements standing out more
from the page.  The tradeoff is in single-voice readability,
i.e. it's easier to read a singe part if the lines are thicker
(especially in poor ligtning conditions, or for people with
poor eyesight), but thinner lines give better overview of the
whole piece (good for conductor).

Anyway, i'm comparing this to a Finale engraving which has
really thin lines.

The problem i'm having is that i cannot change just the staff
thickness - everything else depends on it, so if i make staves
thinner, stems, slurs, hairpins etc. also become thinner, and
i have to compensate for that.